requirements.txt
torch>=1.9.0
torchaudio>=0.9.0
transformers>=4.12.0
librosa>=0.8.1
soundfile>=0.10.3
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
pyyaml>=6.0
wandb>=0.12.0
matplotlib>=3.4.3
tqdm>=4.62.0
flask>=2.0.0
gunicorn>=20.0.0